KYLO-1331 Cloning feed and input processor selection

+++ ui/ui-app/src/main/resources/static/js/feed-mgr/feeds/define-feed/feed-details/DefineFeedDetailsDirective.js
@@ -113,18 +113,23 @@ define(['angular','feed-mgr/feeds/define-feed/module-name'], function (angular,m
          * @param {Object} template the template with properties
          */
         function initializeProperties(template) {
+            if(angular.isDefined(self.model.cloned) && self.model.cloned == true) {
+               var inputProcessors = self.model.inputProcessors;
+                self.inputProcessors = _.sortBy(inputProcessors,'name')
+
+            }else {
                 RegisterTemplateService.initializeProperties(template, 'create', self.model.properties);
                 var inputProcessors = RegisterTemplateService.removeNonUserEditableProperties(template.inputProcessors, true);
                 //sort them by name
-                self.inputProcessors = _.sortBy(inputProcessors,'name')
+                self.inputProcessors = _.sortBy(inputProcessors, 'name')
 
                 self.model.allowPreconditions = template.allowPreconditions;
 
                 self.model.nonInputProcessors = RegisterTemplateService.removeNonUserEditableProperties(template.nonInputProcessors, false);
 
-                if(angular.isDefined(self.model.inputProcessor)){
-                    var match = matchInputProcessor(self.model.inputProcessor,self.inputProcessors);
-                    if(angular.isDefined(match)){
+                if (angular.isDefined(self.model.inputProcessor)) {
+                    var match = matchInputProcessor(self.model.inputProcessor, self.inputProcessors);
+                    if (angular.isDefined(match)) {
                         self.inputProcessor = match;
                         self.inputProcessorId = match.id;
                     }
@@ -134,11 +139,11 @@ define(['angular','feed-mgr/feeds/define-feed/module-name'], function (angular,m
                     self.inputProcessorId = self.inputProcessors[0].id;
                 }
                 // Skip this step if it's empty
-                if (self.inputProcessors.length === 0 && !_.some(self.nonInputProcessors, function(processor) {
+                if (self.inputProcessors.length === 0 && !_.some(self.nonInputProcessors, function (processor) {
                         return processor.userEditable
                     })) {
-                    var step =   StepperService.getStep("DefineFeedStepper", parseInt(self.stepIndex));
-                    if(step != null) {
+                    var step = StepperService.getStep("DefineFeedStepper", parseInt(self.stepIndex));
+                    if (step != null) {
                         step.skip = true;
                     }
                 }
@@ -147,11 +152,11 @@ define(['angular','feed-mgr/feeds/define-feed/module-name'], function (angular,m
                 _.chain(template.inputProcessors.concat(template.nonInputProcessors))
                     .pluck("properties")
                     .flatten(true)
-                    .filter(function(property) {
+                    .filter(function (property) {
                         return angular.isObject(property.propertyDescriptor) && angular.isString(property.propertyDescriptor.identifiesControllerService);
                     })
                     .each(findControllerServicesForProperty);
-
+            }
             self.loading = false;
             validate();
         }
